refactor: migrate helper functions to shared library and update imports
