+++ src/features/admin-panel/categories/Categories.tsx
@@ -2,8 +2,8 @@ import { Group } from '@mantine/core';
 
 import type { Category } from '@/shared/types/category';
 
-import { filterByQuery } from '../helpers';
 import { getCategories } from './services/categories.service';
+import { filterByQuery } from '@/shared/lib/filterByQuery';
 
 export default async function CategoriesPage({ query }: { query: string }) {
     const categories = await getCategories();
