+++ src/shared/lib/getQuery.ts
@@ -1,0 +1,7 @@
+export async function getQuery(searchParams?: Promise<{ query?: string }>) {
+    const resolvedSearchParams = await searchParams;
+
+    const query = resolvedSearchParams?.query || '';
+
+    return query;
+}
